fix: memory64 — widen Limits to u64, fix bulk ops and data segments
Three bugs fixed for memory64 compliance:

1. Limits.min/max widened from u32 to u64 (types.zig)
   Removes the truncation that capped memory64 limits to 4GB.
   Loader now stores full u64 values directly, eliminating the
   raw_min64/raw_max64 workaround. Also propagates is_shared
   from the limits flags byte.

2. Bulk memory ops use i64 for memory64 (interp.zig)
   memory.init, memory.copy, memory.fill now pop dst/src/n as
   i64 when the target memory is memory64, matching the spec.
   Previously hardcoded to i32, causing type mismatches.

3. Data segment offsets support u64 (instance.zig)
   Added evalInitExprAsU64 for memory64 data segments. The
   applyDataSegments function now uses u64 offsets when the
   target memory is memory64, enabling full 64-bit addressing.

Downstream fixes: instance/AOT allocation casts u64 limits to
u32 page counts (capped at 65536), spec test runner updated for
u64 limit comparisons.

Closes #19

+++ src/runtime/common/types.zig
@@ -128,10 +128,10 @@ pub const FuncType = struct {
     pub const Kind = enum(u2) { func, struct_, array };
 };
 
-/// Limits (§2.3.4)
+/// Limits (§2.3.4) — uses u64 to support memory64 proposal.
 pub const Limits = struct {
-    min: u32,
-    max: ?u32 = null,
+    min: u64,
+    max: ?u64 = null,
 };
 
 /// Table type (§2.3.6)
